fix: 폴더 정리
